third_party/WebKit/LayoutTests/http/tests/devtools/elements/shadow/breadcrumb-shadow-roots.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function loaded()
{
    var template = document.querySelector("#tmpl");
    var root = document.querySelector("#host").createShadowRoot();
    root.appendChild(template.content.cloneNode(true));
    var rootClosed = document.querySelector("#hostClosed").attachShadow({mode: 'closed'});
    rootClosed.appendChild(template.content.cloneNode(true));
    runTest();
}

function test() {
  Common.settingForTest('showUAShadowDOM').set(true);
  ElementsTestRunner.expandElementsTree(step0);

  function step0() {
    selectNode(matchUserAgentShadowRoot, step1);
  }

  function step1() {
    ElementsTestRunner.dumpBreadcrumb('User-agent shadow root breadcrumb');
    selectNode(matchOpenShadowRoot, step2);
  }

  function step2() {
    ElementsTestRunner.dumpBreadcrumb('Open shadow root breadcrumb');
    selectNode(matchClosedShadowRoot, step3);
  }

  function step3() {
    ElementsTestRunner.dumpBreadcrumb('Closed shadow root breadcrumb');
    TestRunner.completeTest();
  }

  function selectNode(matchFunction, next) {
    ElementsTestRunner.findNode(matchFunction, callback);
    function callback(node) {
      TestRunner.addSniffer(Elements.ElementsBreadcrumbs.prototype, 'update', next);
      Common.Revealer.revealPromise(node);
    }
  }

  function matchUserAgentShadowRoot(node) {
    return node.shadowRootType() === SDK.DOMNode.ShadowRootTypes.UserAgent;
  }

  function matchOpenShadowRoot(node) {
    return node.shadowRootType() === SDK.DOMNode.ShadowRootTypes.Open;
  }

  function matchClosedShadowRoot(node) {
    return node.shadowRootType() === SDK.DOMNode.ShadowRootTypes.Closed;
  }
}

</script>
</head>

<body onload="loaded()">
<p>
Tests that shadow roots are displayed correctly in breadcrumbs.
</p>

<input type="text">
<div id="host"></div>
<div id="hostClosed"></div>
<template id="tmpl">
    <style>.red { color: red; }</style>
    <div id="inner" class="red">inner</div>
</template>


</body>
</html>
