third_party/WebKit/LayoutTests/http/tests/devtools/elements/shadow/breadcrumb-shadow-roots-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that shadow roots are displayed correctly in breadcrumbs.

User-agent shadow root breadcrumb:
html > body > input > #shadow-root
Open shadow root breadcrumb:
html > body > div#host > #shadow-root
Closed shadow root breadcrumb:
html > body > div#hostClosed > #shadow-root
